requirements_txt.txt
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
matplotlib==3.7.2
plotly==5.15.0
streamlit==1.28.1
pyyaml==6.0.1